mod6_labs/requirements.txt
flet==0.28.3
httpx==0.27.0
python-dotenv==1.0.0

